Add exceptions module.
