+++ futaba/__init__.py
@@ -14,20 +14,4 @@
 futaba - A Discord Mod bot for the Programming server
 '''
 
-from . import client, config, download, enums, journal, parse, permissions, str_builder, unicode, utils
-
-__all__ = [
-    '__version__',
-    'client',
-    'config',
-    'download',
-    'enums',
-    'journal',
-    'parse',
-    'permissions',
-    'str_builder',
-    'unicode',
-    'utils',
-]
-
-__version__ = '0.1.1'
+__version__ = '0.1.2'
